Refactor: Extract DiscoverStructure.ts into 7 focused modules (#1472)
Decompose DiscoverStructure.ts (3,857 → 1,823 lines) into single-responsibility
modules: DiscoverLogging, FocusSelection, DiscoverSquashAnalysis, DiscoverAnalysis,
NeuronImpact, DiscoverDataLoading, and RustAnalysisCache. All 3,823 tests pass
with no functional changes.
